[ADD] Repas a deux tarifs, journee de travail, et correctifs
Deux modules seulement : his_meal_management et maintenance_university.

REPAS — six forfaits, deux repas, un seul portefeuille
Le tarif du groupe a deux colonnes. Un credit vaut un repas a 600 DA, un repas
a 300 DA en coute la moitie. Les deux colonnes valorisent le credit a
l'identique a duree egale (500 / 480 / 450 DA), donc rien a arbitrer entre
elles. Les credits deviennent decimaux - tout montant est un multiple de 0,5,
exact en binaire - et ne perimeent plus : date_end devient facultatif.

meal_credit_cost sur le produit remplace meal_product_id sur pos.config. Un
repas est un produit qui coute des credits, pas un champ d'une caisse : toutes
les caisses servent tous les repas sans rien configurer, ce qui permet enfin a
la Cafeteria d'en servir un. Le post_init_hook qui cablait le Restaurant
disparait avec le champ. _consume_meal_credit preleve un montant decimal sur
plusieurs abonnements si besoin.

MAINTENANCE — la journee de travail
maintenance.university.workday enregistre l'arrivee, chaque pause et le depart,
avec une banniere en tete de My Work (Start / Pause / Reprise / Fin). C'est la
presence, distincte du temps impute aux demandes : l'ecart est le trajet et
l'attente, et le Recap affiche les deux. Le travailleur est en lecture seule
sur le modele et les boutons ecrivent sous sudo() apres verification de la
propriete - un releve que la personne mesuree peut editer ne mesure rien. Rien
ne se ferme tout seul : l'ecran Work Days sert aux corrections.

Le chef voit qui travaille et qui est en pause depuis Worker Summary, avec
l'historique des journees et de leurs segments.

CORRECTIFS
- action_convert_to_request ne nommait aucune vue : maintenance.request porte
  deux formulaires de meme priorite et la resolution retombait sur celui du
  coeur, sans Batiment ni Categorie, tous deux obligatoires.
- My Work etait devenu une action client embarquant un <View> nu : createRecord
  et selectRecord sont optionnelles sans defaut, donc Nouveau et l'ouverture
  d'une carte ne faisaient rien. Retour a une action fenetre, banniere par
  js_class.
- Les boutons repas affichaient un libelle vide : la variante POS ne delegue au
  modele que ses methodes, jamais un champ non charge.

100 tests, 0 echec, installes a neuf sur cette branche.

+++ his_meal_management/models/meal_card.py
@@ -48,7 +48,7 @@ class HisMealCard(models.Model):
     replaced_card_id = fields.Many2one(
         'his.meal.card', string="Replaces", ondelete='set null',
     )
-    credits_remaining = fields.Integer(
+    credits_remaining = fields.Float(
         related='partner_id.meal_credits_remaining', string="Credits", readonly=True,
     )
